rename the plugin file to webpack plugins
