+++ packages/craco/lib/webpack-plugins.js
